+++ CHANGELOG.md
@@ -14,11 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### **Changed**
 
 - updated object detection to YOLOv8+
-- made `modules/optionals/datalake-buckets` module generic
+- made `modules/optionals/datalake-buckets` module generic and bumped CDK version
 - made the following module parameters generic:
   - `examples/example-dags`
   - `examples/example-tf`
   - `examples/example-tf-prereqs`
+- made `modules/core/metadata-storage` module generic and bumped CDK version
 
 ### **Removed**